+++ doc/source/oneview_redfish_toolkit.blueprints.rst
@@ -12,6 +12,14 @@ oneview\_redfish\_toolkit\.blueprints\.computer\_system\_collection module
     :undoc-members:
     :show-inheritance:
 
+oneview\_redfish\_toolkit\.blueprints\.redfish\_base module
+-----------------------------------------------------------
+
+.. automodule:: oneview_redfish_toolkit.blueprints.redfish_base
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 oneview\_redfish\_toolkit\.blueprints\.service\_root module
 -----------------------------------------------------------
