+++ oneview_redfish_toolkit/tests/mockups/ServerHardwares.json
