+++ oneview_redfish_toolkit/tests/test_util_load_schemas.py
